requirements.txt
python-dotenv
torch
sentence-transformers
langchain
openai
youtube-transcript-api
faiss-cpu
streamlit